Split LocalAssetsManager to abstract and imp
This is only a part of a bigger change.
