+++ playkit/src/main/java/com/kaltura/playkit/LocalAssetsManager.java
@@ -33,6 +33,8 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+
 
 /**
  * Responsible for managing the local(offline) assets. When offline playback of the
@@ -40,20 +42,11 @@ import java.util.List;
  * Note, you must have network connection, while register.
  * Created by anton.afanasiev on 13/12/2016.
  */
-public class LocalAssetsManager {
+public abstract class LocalAssetsManager {
 
-    private static final PKLog log = PKLog.get("LocalAssetsManager");
-    private static final String ASSET_ID_PREFIX = "assetId:";
+    static final PKLog log = PKLog.get("LocalAssetsManager");
 
-    private final Context context;
-    private LocalDataStore localDataStore;
-    private PKRequestParams.Adapter licenseRequestParamAdapter;
-
-    private Handler mainHandler = new Handler(Looper.getMainLooper());
-
-    public void setLicenseRequestAdapter(PKRequestParams.Adapter licenseRequestAdapter) {
-        this.licenseRequestParamAdapter = licenseRequestAdapter;
-    }
+    public abstract void setLicenseRequestAdapter(PKRequestParams.Adapter licenseRequestAdapter);
 
     /**
      * Listener that notify about the result when registration flow is ended.
@@ -90,28 +83,6 @@ public class LocalAssetsManager {
     }
 
     /**
-     * Constructor which will create {@link DefaultLocalDataStore}
-     *
-     * @param context - the application context.
-     */
-    public LocalAssetsManager(Context context) {
-        this(context, new DefaultLocalDataStore(context));
-    }
-
-    /**
-     * Constructor with custom implementation of the {@link LocalDataStore}
-     *
-     * @param context        - the application context.
-     * @param localDataStore - custom implementation of {@link LocalDataStore}
-     */
-    public LocalAssetsManager(Context context, LocalDataStore localDataStore) {
-        this.context = context;
-        this.localDataStore = localDataStore;
-
-        MediaSupport.initializeDrm(context, null);
-    }
-
-    /**
      * Register the asset. If the asset have drm protection it will store keySetId and {@link PKMediaFormat} in {@link LocalDataStore}
      * If no drm available only {@link PKMediaFormat as byte[]} will be stored.
      *
@@ -120,79 +91,8 @@ public class LocalAssetsManager {
      * @param assetId        - the asset id.
      * @param listener       - notify about the success/fail after the completion of the registration process.
      */
-    public void registerAsset(@NonNull final PKMediaSource mediaSource, @NonNull final String localAssetPath,
-                              @NonNull final String assetId, final AssetRegistrationListener listener) {
-
-        checkIfParamsAreValid(mediaSource.getUrl(), localAssetPath, assetId);
-
-        if (!isOnline(context)) {
-            listener.onFailed(localAssetPath, new Exception("Can't register/refresh when offline"));
-            return;
-        }
-
-        PKDrmParams drmParams = findSupportedDrmParams(mediaSource);
-
-        PKMediaFormat mediaFormat = mediaSource.getMediaFormat();
-        if (mediaFormat == null) {
-            listener.onFailed(localAssetPath,
-                    new IllegalArgumentException("Can not register media, when PKMediaFormat and url of PKMediaSource not exist."));
-        }
-
-        if (drmParams != null) {
-            registerDrmAsset(localAssetPath, assetId, mediaFormat, drmParams, listener);
-        } else {
-            registerClearAsset(localAssetPath, assetId, mediaFormat, listener);
-        }
-    }
-
-    public void registerDrmAsset(byte[] drmInitData, String mimeType, final String assetId, final PKMediaFormat mediaFormat, String licenseUri, final AssetRegistrationListener listener) {
-        doInBackground(() -> {
-            final WidevineModularAdapter widevine = new WidevineModularAdapter(context, localDataStore);
-            widevine.registerAsset(drmInitData, mimeType, licenseUri, licenseRequestParamAdapter, listener);
-            localDataStore.save(buildAssetKey(assetId), buildMediaFormatValueAsByteArray(mediaFormat, PKDrmParams.Scheme.WidevineCENC));
-        });
-    }
-
-    /**
-     * Will register the drm asset and store the keyset id and {@link PKMediaFormat} in local storage.
-     *
-     * @param localAssetPath - the local asset path of the asset.
-     * @param assetId        - the asset id.
-     * @param mediaFormat    - the media format converted to byte[].
-     * @param drmParams      - drm params of the media.
-     * @param listener       - notify about the success/fail after the completion of the registration process.
-     */
-    private void registerDrmAsset(final String localAssetPath, final String assetId, final PKMediaFormat mediaFormat, final PKDrmParams drmParams, final AssetRegistrationListener listener) {
-        doInBackground(() -> {
-            try {
-                DrmAdapter drmAdapter = DrmAdapter.getDrmAdapter(drmParams.getScheme(), context, localDataStore);
-                String licenseUri = drmParams.getLicenseUri();
-
-                boolean isRegistered = drmAdapter.registerAsset(localAssetPath, assetId, licenseUri, licenseRequestParamAdapter, listener);
-                if (isRegistered) {
-                    localDataStore.save(buildAssetKey(assetId), buildMediaFormatValueAsByteArray(mediaFormat, drmParams.getScheme()));
-                }
-            } catch (final IOException e) {
-                log.e("Error", e);
-                if (listener != null) {
-                    mainHandler.post(() -> listener.onFailed(localAssetPath, e));
-                }
-            }
-        });
-    }
-
-    /**
-     * Will register clear asset and store {@link PKMediaFormat} in local storage.
-     *
-     * @param localAssetPath - the local asset path of the asset.
-     * @param assetId        - the asset id.
-     * @param mediaFormat    - the media format converted to byte[].
-     * @param listener       - notify about the success/fail after the completion of the registration process.
-     */
-    private void registerClearAsset(String localAssetPath, String assetId, PKMediaFormat mediaFormat, AssetRegistrationListener listener) {
-        localDataStore.save(buildAssetKey(assetId), buildMediaFormatValueAsByteArray(mediaFormat, null));
-        listener.onRegistered(localAssetPath);
-    }
+    public abstract void registerAsset(@NonNull PKMediaSource mediaSource, @NonNull String localAssetPath,
+                                       @NonNull String assetId, AssetRegistrationListener listener);
 
     /**
      * Unregister asset. If the asset have drm protection it will be removed from {@link LocalDataStore}
@@ -202,37 +102,11 @@ public class LocalAssetsManager {
      * @param assetId        - the asset id
      * @param listener       - notify when the asset is removed.
      */
-    public void unregisterAsset(@NonNull final String localAssetPath,
-                                @NonNull final String assetId, final AssetRemovalListener listener) {
+    public abstract void unregisterAsset(@NonNull String localAssetPath,
+                                         @NonNull String assetId, AssetRemovalListener listener);
 
-
-        PKDrmParams.Scheme scheme = getLocalAssetScheme(assetId, localDataStore);
-
-        if (scheme == null) {
-            removeAsset(localAssetPath, assetId, listener);
-            return;
-        }
-
-        final DrmAdapter drmAdapter = DrmAdapter.getDrmAdapter(scheme, context, localDataStore);
-
-        doInBackground(() -> {
-            drmAdapter.unregisterAsset(localAssetPath, assetId, localAssetPath1 -> {
-                mainHandler.post(() -> {
-                    removeAsset(localAssetPath1, assetId, listener);
-                });
-            });
-        });
-    }
-
-    public void refreshAsset(@NonNull final PKMediaSource mediaSource, @NonNull final String localAssetPath,
-                             @NonNull final String assetId, final AssetRegistrationListener listener) {
-        registerAsset(mediaSource, localAssetPath, assetId, listener);
-    }
-
-    private void removeAsset(String localAssetPath, String assetId, AssetRemovalListener listener) {
-        localDataStore.remove(buildAssetKey(assetId));
-        listener.onRemoved(localAssetPath);
-    }
+    public abstract void refreshAsset(@NonNull PKMediaSource mediaSource, @NonNull String localAssetPath,
+                                      @NonNull String assetId, AssetRegistrationListener listener);
 
 
     /**
@@ -242,210 +116,35 @@ public class LocalAssetsManager {
      * @param assetId        - the asset id.
      * @param listener       - will pass the result of the status.
      */
-    public void checkAssetStatus(@NonNull final String localAssetPath, @NonNull final String assetId,
-                                 @Nullable final AssetStatusListener listener) {
-
-        PKDrmParams.Scheme scheme = getLocalAssetScheme(assetId, localDataStore);
-        if (scheme == null) {
-            checkClearAssetStatus(localAssetPath, assetId, listener);
-        } else {
-            checkDrmAssetStatus(localAssetPath, assetId, scheme, listener);
-        }
-    }
-
-    private void checkDrmAssetStatus(final String localAssetPath, final String assetId, PKDrmParams.Scheme scheme, final AssetStatusListener listener) {
-
-        final DrmAdapter drmAdapter = DrmAdapter.getDrmAdapter(scheme, context, localDataStore);
-
-        doInBackground(() -> drmAdapter.checkAssetStatus(localAssetPath, assetId, (localAssetPath1, expiryTimeSeconds, availableTimeSeconds, isRegistered) -> {
-            if (listener != null) {
-                mainHandler.post(() ->  {
-                    listener.onStatus(localAssetPath1, expiryTimeSeconds, availableTimeSeconds, isRegistered);
-                });
-            }
-        }));
-    }
-
-    private void checkClearAssetStatus(String localAssetPath, String assetId, AssetStatusListener listener) {
-        try {
-            localDataStore.load(buildAssetKey(assetId));
-            listener.onStatus(localAssetPath, Long.MAX_VALUE, Long.MAX_VALUE, true);
-        } catch (FileNotFoundException e) {
-            listener.onStatus(localAssetPath, 0, 0, false);
-        }
-    }
-
-    private static PKDrmParams findSupportedDrmParams(@NonNull PKMediaSource mediaSource) {
-        if (mediaSource.getDrmData() == null) {
-            return null;
-        }
-
-        for (PKDrmParams params : mediaSource.getDrmData()) {
-            switch (params.getScheme()) {
-                case WidevineCENC:
-                    if (MediaSupport.widevineModular()) {
-                        return params;
-                    }
-                    break;
-                case WidevineClassic:
-                    if (MediaSupport.widevineClassic()) {
-                        return params;
-                    }
-                    break;
-                case PlayReadyCENC:
-                    log.d("Skipping unsupported PlayReady params");
-                    break;
-            }
-        }
-        return null;
-    }
-
-    @Nullable
-    private static PKDrmParams.Scheme getLocalAssetScheme(@NonNull String assetId, LocalDataStore localDataStore) {
-        try {
-            String mediaFormatValue = new String(localDataStore.load(buildAssetKey(assetId)));
-            String[] splitFormatValue = mediaFormatValue.split(":");
-            String schemeName = splitFormatValue[1];
-            if (schemeName.equals("null")) {
-                return null;
-            }
-
-            return PKDrmParams.Scheme.valueOf(schemeName);
-
-        } catch (FileNotFoundException e) {
-            log.e(e.getMessage());
-            return null;
-        }
-    }
+    public abstract void checkAssetStatus(@NonNull String localAssetPath, @NonNull String assetId,
+                                          @Nullable AssetStatusListener listener);
 
     /**
      * @param assetId        - the id of the asset.
      * @param localAssetPath - the actual url of the video that should be played.
      * @return - the {@link PKMediaSource} that should be passed to the player.
      */
-    public PKMediaSource getLocalMediaSource(@NonNull final String assetId, @NonNull final String localAssetPath) {
-        return new LocalMediaSource(localDataStore, localAssetPath, assetId);
-    }
-
-    /**
-     * @param assetId        - the id of the asset.
-//     * @param exoMediaSource - the actual url of the video that should be played.
-     * @return - the {@link PKMediaSource} that should be passed to the player.
-     */
-    public PKMediaSource getLocalMediaSource(@NonNull final String assetId, @NonNull final MediaSource exoMediaSource) {
-        return new LocalExoMediaSource(localDataStore, exoMediaSource, assetId);
-    }
-
-    /**
-     * Will check if passed parameters are valid.
-     *
-     * @param url            - url of the media.
-     * @param localAssetPath - the local asset path of the media.
-     * @param assetId        - the asset id.
-     */
-    private void checkIfParamsAreValid(String url, String localAssetPath, String assetId) {
-        checkNotEmpty(url, "mediaSource.url");
-        checkNotEmpty(localAssetPath, "localAssetPath");
-        checkNotEmpty(assetId, "assetId");
-    }
-
-    /**
-     * Checking arguments. if the argument is not valid, will throw an {@link IllegalArgumentException}
-     *
-     * @param invalid - the state of the argument.
-     * @param message - message to print, to the console.
-     */
-    private void checkArg(boolean invalid, String message) {
-        if (invalid) {
-            throw new IllegalArgumentException(message);
-        }
-    }
-
-    /**
-     * check the passed String.
-     *
-     * @param str  - String to check.
-     * @param name - the descriptive name of the String.
-     */
-    private void checkNotEmpty(String str, String name) {
-        checkArg(TextUtils.isEmpty(str), name + " must not be empty");
-    }
-
-    private void doInBackground(Runnable runnable) {
-        AsyncTask.execute(runnable);
-    }
-
-    /**
-     * Check if network connection is available.
-     *
-     * @param context - context.
-     * @return - true if there is network connection, otherwise - false.
-     */
-    private boolean isOnline(Context context) {
-        ConnectivityManager conMgr = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo netInfo = conMgr.getActiveNetworkInfo();
-        return !(netInfo == null || !netInfo.isConnected() || !netInfo.isAvailable());
-    }
-
-    private static String buildAssetKey(String assetId) {
-        return ASSET_ID_PREFIX + assetId;
-    }
-
-    private byte[] buildMediaFormatValueAsByteArray(PKMediaFormat mediaFormat, PKDrmParams.Scheme scheme) {
-        String mediaFormatName = mediaFormat.toString();
-        String schemeName = "null";
-        if (scheme != null) {
-            schemeName = scheme.toString();
-        }
-        String stringBuilder = mediaFormatName +
-                ":" +
-                schemeName;
-        return stringBuilder.getBytes();
-    }
+    public abstract PKMediaSource getLocalMediaSource(@NonNull String assetId, @NonNull String localAssetPath);
 
     /**
      * The local media source that should be passed to the player
      * when offline(locally stored) media want to be played.
      */
-    public static class LocalMediaSource extends PKMediaSource {
-
-        private PKDrmParams.Scheme scheme;
-        private LocalDataStore localDataStore;
-
-
-
-        /**
-         * @param localDataStore - the storage from where drm keySetId is stored.
-         * @param localPath      - the local url of the media.
-         * @param assetId        - the id of the media.
-         */
-        LocalMediaSource(LocalDataStore localDataStore, String localPath, String assetId) {
-            setId(assetId);
-            setUrl(localPath);
-            this.localDataStore = localDataStore;
-
-            this.scheme = getLocalAssetScheme(assetId, localDataStore);
-        }
+    public static abstract class LocalMediaSource extends PKMediaSource {
 
         /**
          * @return - the {@link LocalDataStore}
          */
-        public LocalDataStore getStorage() {
-            return localDataStore;
-        }
+        public abstract LocalDataStore getStorage();
 
         @Override
-        public boolean hasDrmParams() {
-            return this.scheme != null;
-        }
+        public abstract boolean hasDrmParams();
 
         @Override
-        public List<PKDrmParams> getDrmData() {
-            return Collections.emptyList();
-        }
+        public abstract List<PKDrmParams> getDrmData();
     }
 
-    public static class LocalExoMediaSource extends LocalMediaSource {
+    public static class LocalExoMediaSource extends LocalAssetsManagerImp.LocalMediaSourceImp {
         private MediaSource exoMediaSource;
 
         /**
@@ -461,7 +160,6 @@ public class LocalAssetsManager {
         public MediaSource getExoMediaSource() {
             return exoMediaSource;
         }
-
     }
 
     /**
@@ -510,6 +208,50 @@ public class LocalAssetsManager {
             sharedPreferences.edit()
                     .remove(key)
                     .apply();
+        }
+    }
+
+    public static class AssetStatus {
+        public final boolean registered;
+        public final long licenseDuration;
+        public final long totalDuration;
+        public final boolean hasContentProtection;
+
+        private AssetStatus(boolean registered, long licenseDuration, long totalDuration, boolean hasContentProtection) {
+            this.registered = registered;
+            this.licenseDuration = licenseDuration;
+            this.totalDuration = totalDuration;
+            this.hasContentProtection = hasContentProtection;
+        }
+
+        public static AssetStatus fromWidevineMap(Map<String, String> map) {
+            long licenseDurationRemaining = 0;
+            long playbackDurationRemaining = 0;
+            try {
+                final String licenseDurationRemainingString = map.get("LicenseDurationRemaining");
+                final String playbackDurationRemainingString = map.get("PlaybackDurationRemaining");
+                if (playbackDurationRemainingString == null || licenseDurationRemainingString == null) {
+                    log.e("Missing keys in KeyStatus: " + map);
+                    return withDrm(false, -1, -1);
+                }
+
+                licenseDurationRemaining = Long.parseLong(licenseDurationRemainingString);
+                playbackDurationRemaining = Long.parseLong(playbackDurationRemainingString);
+
+                return withDrm(true, licenseDurationRemaining, playbackDurationRemaining);
+
+            } catch (NumberFormatException e) {
+                log.e("Invalid integers in KeyStatus: " + map);
+                return withDrm(false, -1, -1);
+            }
+        }
+
+        public static AssetStatus clear(boolean registered) {
+            return new AssetStatus(registered, 0, 0, false);
+        }
+
+        public static AssetStatus withDrm(boolean registered, long licenseDuration, long totalDuration) {
+            return new AssetStatus(registered, licenseDuration, totalDuration, true);
         }
     }
