+++ playkit/src/main/java/com/kaltura/playkit/drm/WidevineModularAdapter.java
@@ -32,11 +32,7 @@ import com.kaltura.android.exoplayer2.drm.HttpMediaDrmCallback;
 import com.kaltura.android.exoplayer2.drm.UnsupportedDrmException;
 import com.kaltura.android.exoplayer2.upstream.DefaultHttpDataSourceFactory;
 import com.kaltura.android.exoplayer2.upstream.HttpDataSource;
-import com.kaltura.playkit.BuildConfig;
-import com.kaltura.playkit.LocalAssetsManager;
-import com.kaltura.playkit.LocalDataStore;
-import com.kaltura.playkit.PKLog;
-import com.kaltura.playkit.PKRequestParams;
+import com.kaltura.playkit.*;
 import com.kaltura.playkit.player.MediaSupport;
 
 import java.io.FileNotFoundException;
@@ -73,7 +69,7 @@ public class WidevineModularAdapter extends DrmAdapter {
                 listener.onRegistered(localAssetPath);
             }
             return result;
-        } catch (RegisterException e) {
+        } catch (LocalAssetsManagerImp.RegisterException e) {
             if (listener != null) {
                 listener.onFailed(localAssetPath, e);
             }
@@ -81,7 +77,7 @@ public class WidevineModularAdapter extends DrmAdapter {
         }
     }
 
-    private boolean registerAsset(String localAssetPath, String assetId, String licenseUri, PKRequestParams.Adapter requestParamsAdapter) throws RegisterException {
+    private boolean registerAsset(String localAssetPath, String assetId, String licenseUri, PKRequestParams.Adapter requestParamsAdapter) throws LocalAssetsManagerImp.RegisterException {
 
         // obtain the dash manifest.
         SimpleDashParser dash = parseDash(localAssetPath, assetId);
@@ -94,32 +90,23 @@ public class WidevineModularAdapter extends DrmAdapter {
         String mimeType = dash.format.containerMimeType;
         byte[] initData = dash.widevineInitData;
 
-        registerAsset(licenseUri, requestParamsAdapter, mimeType, initData);
+        registerAsset(initData, mimeType, licenseUri, requestParamsAdapter);
 
         return true;
     }
 
-    private void registerAsset(String licenseUri, PKRequestParams.Adapter requestParamsAdapter, String mimeType, byte[] initData) throws RegisterException {
+    public void registerAsset(byte[] initData, String mimeType, String licenseUri, PKRequestParams.Adapter requestParamsAdapter) throws LocalAssetsManagerImp.RegisterException {
         final byte[] offlineKeyId = downloadOfflineLicense(licenseUri, requestParamsAdapter, mimeType, initData);
         localDataStore.save(toBase64(initData), offlineKeyId);
     }
 
-    public void registerAsset(byte[] initData, String mimeType, String licenseUri, PKRequestParams.Adapter requestParamsAdapter, LocalAssetsManager.AssetRegistrationListener listener) {
-        try {
-            registerAsset(licenseUri, requestParamsAdapter, mimeType, initData);
-            listener.onRegistered(null);
-        } catch (RegisterException e) {
-            listener.onFailed(null, e);
-        }
-    }
-
-    private byte[] downloadOfflineLicense(String licenseUri, PKRequestParams.Adapter requestParamsAdapter, String mimeType, byte[] initData) throws RegisterException {
+    private byte[] downloadOfflineLicense(String licenseUri, PKRequestParams.Adapter requestParamsAdapter, String mimeType, byte[] initData) throws LocalAssetsManagerImp.RegisterException {
         MediaDrmSession session;
         FrameworkMediaDrm mediaDrm = createMediaDrm();
         try {
             session = MediaDrmSession.open(mediaDrm);
         } catch (MediaDrmException e) {
-            throw new RegisterException("Can't open session", e);
+            throw new LocalAssetsManagerImp.RegisterException("Can't open session", e);
         }
 
         // Get keyRequest
@@ -137,7 +124,7 @@ public class WidevineModularAdapter extends DrmAdapter {
                 log.d("registerAsset: response data (b64): " + toBase64(keyResponse));
 
             } catch (Exception e) {
-                throw new RegisterException("Can't send key request for registration", e);
+                throw new LocalAssetsManagerImp.RegisterException("Can't send key request for registration", e);
             }
 
             // Provide keyResponse
@@ -145,11 +132,11 @@ public class WidevineModularAdapter extends DrmAdapter {
                 return session.provideKeyResponse(keyResponse);
 
             } catch (DeniedByServerException e) {
-                throw new RegisterException("Request denied by server", e);
+                throw new LocalAssetsManagerImp.RegisterException("Request denied by server", e);
             }
 
         } catch (WidevineNotSupportedException e) {
-            throw new RegisterException("Can't execute KeyRequest", e);
+            throw new LocalAssetsManagerImp.RegisterException("Can't execute KeyRequest", e);
 
         } finally {
             session.close();
@@ -162,7 +149,7 @@ public class WidevineModularAdapter extends DrmAdapter {
         try {
             unregisterAsset(localAssetPath, assetId);
             return true;
-        } catch (RegisterException e) {
+        } catch (LocalAssetsManagerImp.RegisterException e) {
             log.e("Failed to unregister", e);
             return false;
         } finally {
@@ -172,7 +159,7 @@ public class WidevineModularAdapter extends DrmAdapter {
         }
     }
 
-    private boolean unregisterAsset(String localAssetPath, String assetId) throws RegisterException {
+    private boolean unregisterAsset(String localAssetPath, String assetId) throws LocalAssetsManagerImp.RegisterException {
 
         SimpleDashParser dash = parseDash(localAssetPath, assetId);
         if (!dash.hasContentProtection) {
@@ -187,7 +174,7 @@ public class WidevineModularAdapter extends DrmAdapter {
         try {
             keySetId = localDataStore.load(key);
         } catch (FileNotFoundException e) {
-            throw new RegisterException("Can't unregister -- keySetId not found", e);
+            throw new LocalAssetsManagerImp.RegisterException("Can't unregister -- keySetId not found", e);
         }
 
         FrameworkMediaDrm mediaDrm = createMediaDrm();
@@ -241,7 +228,7 @@ public class WidevineModularAdapter extends DrmAdapter {
                 listener.onStatus(localAssetPath, -1, -1, false);
             }
             return false;
-        } catch (RegisterException e) {
+        } catch (LocalAssetsManagerImp.RegisterException e) {
             if (listener != null) {
                 listener.onStatus(localAssetPath, 0, 0, false);
             }
@@ -256,7 +243,7 @@ public class WidevineModularAdapter extends DrmAdapter {
         return true;
     }
 
-    private Map<String, String> checkAssetStatus(String localAssetPath, String assetId) throws RegisterException {
+    private Map<String, String> checkAssetStatus(String localAssetPath, String assetId) throws LocalAssetsManagerImp.RegisterException {
         SimpleDashParser dash = parseDash(localAssetPath, assetId);
 
         //no content protection, so there could not be any status info, so return null.
@@ -275,7 +262,7 @@ public class WidevineModularAdapter extends DrmAdapter {
             String key = toBase64(dash.widevineInitData);
             session = openSessionWithKeys(mediaDrm, key);
         } catch (MediaDrmException | FileNotFoundException | MediaCryptoException e) {
-            throw new RegisterException("Can't open session with keys", e);
+            throw new LocalAssetsManagerImp.RegisterException("Can't open session with keys", e);
         }
 
 
@@ -289,7 +276,7 @@ public class WidevineModularAdapter extends DrmAdapter {
     }
 
     @NonNull
-    private FrameworkMediaDrm createMediaDrm() throws RegisterException {
+    private FrameworkMediaDrm createMediaDrm() throws LocalAssetsManagerImp.RegisterException {
         FrameworkMediaDrm mediaDrm = null;
         try {
             mediaDrm = FrameworkMediaDrm.newInstance(MediaSupport.WIDEVINE_UUID);
@@ -298,7 +285,7 @@ public class WidevineModularAdapter extends DrmAdapter {
         }
 
         if (mediaDrm == null) {
-            throw new RegisterException("Could not create MediaDrm instance ", null);
+            throw new LocalAssetsManagerImp.RegisterException("Could not create MediaDrm instance ", null);
         }
 
         return mediaDrm;
@@ -310,20 +297,20 @@ public class WidevineModularAdapter extends DrmAdapter {
      * @param localPath - file from which to parse the dash manifest.
      * @param assetId   - the asset id.
      * @return - {@link SimpleDashParser} which contains the manifest data we need.
-     * @throws RegisterException - {@link RegisterException}
+     * @throws LocalAssetsManagerImp.RegisterException - {@link LocalAssetsManagerImp.RegisterException}
      */
-    private SimpleDashParser parseDash(String localPath, String assetId) throws RegisterException {
+    private SimpleDashParser parseDash(String localPath, String assetId) throws LocalAssetsManagerImp.RegisterException {
         SimpleDashParser dashParser;
         try {
             dashParser = new SimpleDashParser().parse(localPath, assetId);
             if (dashParser.format == null) {
-                throw new RegisterException("Unknown format", null);
+                throw new LocalAssetsManagerImp.RegisterException("Unknown format", null);
             }
             if (dashParser.hasContentProtection && dashParser.widevineInitData == null) {
                 throw new NoWidevinePSSHException("No Widevine PSSH in media", null);
             }
         } catch (IOException e) {
-            throw new RegisterException("Can't parse local dash", e);
+            throw new LocalAssetsManagerImp.RegisterException("Can't parse local dash", e);
         }
 
         return dashParser;
@@ -378,13 +365,7 @@ public class WidevineModularAdapter extends DrmAdapter {
                 + ") " + "ExoPlayerLib/" + ExoPlayerLibraryInfo.VERSION;
     }
 
-    private class RegisterException extends Exception {
-        RegisterException(String detailMessage, Throwable throwable) {
-            super(detailMessage, throwable);
-        }
-    }
-
-    private class NoWidevinePSSHException extends RegisterException {
+    private class NoWidevinePSSHException extends LocalAssetsManagerImp.RegisterException {
         NoWidevinePSSHException(String detailMessage, Throwable throwable) {
             super(detailMessage, throwable);
         }
